Upgrade scala and play
