+++ app/controllers/predicates/AuthorisedAction.scala
@@ -52,7 +52,7 @@ class AuthorisedAction @Inject() ()(implicit
           logger.warn("[AuthorisedAction][async] - No MTDITID in the header. Returning unauthorised.")
           unauthorized
         }(mtdItId =>
-          authorised.retrieve(affinityGroup) {
+          authorised().retrieve(affinityGroup) {
             case Some(AffinityGroup.Agent) => agentAuthentication(block, mtdItId)(request, headerCarrier)
             case _                         => individualAuthentication(block, mtdItId)(request, headerCarrier)
           } recover {
@@ -70,7 +70,7 @@ class AuthorisedAction @Inject() ()(implicit
     request: Request[A],
     hc: HeaderCarrier
   ): Future[Result] =
-    authorised.retrieve(allEnrolments and confidenceLevel) {
+    authorised().retrieve(allEnrolments and confidenceLevel) {
       case enrolments ~ userConfidence if userConfidence.level >= minimumConfidenceLevel =>
         val optionalMtdItId: Option[String] =
           enrolmentGetIdentifierValue(EnrolmentKeys.Individual, EnrolmentIdentifiers.individualId, enrolments)
