+++ it/test/api/IgnoreStateBenefitITest.scala
@@ -14,29 +14,24 @@
  * limitations under the License.
  */
 
-package api
+package test.api
 
 import com.github.tomakehurst.wiremock.http.HttpHeader
-import helpers.{AuthStub, WiremockSpec}
-import models.{CreateUpdateOverrideStateBenefit, DesErrorBodyModel}
+import models.{DesErrorBodyModel, IgnoreStateBenefit}
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.time.{Seconds, Span}
 import org.scalatestplus.play.PlaySpec
 import play.api.http.HeaderNames
 import play.api.http.Status._
 import play.api.libs.json.{JsValue, Json}
+import test.helpers.{AuthStub, WiremockSpec}
 import utils.DESTaxYearHelper.desTaxYearConverter
 
-class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec with ScalaFutures with AuthStub {
+class IgnoreStateBenefitITest extends PlaySpec with WiremockSpec with ScalaFutures with AuthStub {
 
-  val fullCreateUpdateStateBenefitData: CreateUpdateOverrideStateBenefit = CreateUpdateOverrideStateBenefit(
-    amount = 21.22, taxPaid = Some(0.50))
+  val fullIgnoreStateBenefit: IgnoreStateBenefit = IgnoreStateBenefit(true)
 
-  val fullCreateUpdateStateBenefitJson: JsValue = Json.parse(
-    """{
-      |	"amount": 21.22,
-      |	"taxPaid": 0.50
-      |}""".stripMargin)
+  val fullCreateUpdateStateBenefitJson: JsValue = Json.parse("""{"ignoreBenefit": true}""")
 
   trait Setup {
     val timeSpan: Long = 5
@@ -47,16 +42,16 @@ class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec w
     val authorization: (String, String) = HeaderNames.AUTHORIZATION -> "mock-bearer-token"
     val benefitId: String = "a111111a-abcd-111a-123a-11a1a111a1"
     val requestHeaders: Seq[HttpHeader] = Seq(new HttpHeader("mtditid", "555555555"))
-    val desUrl = s"/income-tax/income/state-benefits/$nino/${desTaxYearConverter(taxYear)}/$benefitId"
-    val serviceUrl: String = s"/income-tax-benefits/state-benefits/override/nino/$nino/taxYear/$taxYear/benefitId/$benefitId"
+    val desUrl = s"/income-tax/income/state-benefits/$nino/${desTaxYearConverter(taxYear)}/ignore/$benefitId"
+    val serviceUrl: String = s"/income-tax-benefits/state-benefits/nino/$nino/taxYear/$taxYear/benefitId/$benefitId/ignoreBenefit/true"
 
     auditStubs()
   }
 
-  "create or update state benefit" when {
+  "ignore state benefit" should {
     "return a No content Success response" in new Setup {
 
-      stubPutWithoutResponseBody(desUrl, Json.toJson(fullCreateUpdateStateBenefitData).toString(), NO_CONTENT)
+      stubPutWithoutResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), NO_CONTENT)
 
       authorised()
 
@@ -69,24 +64,12 @@ class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec w
       }
     }
 
-    "return 400 if the body payload validation fails" in new Setup {
-
-      authorised()
-
-      whenReady(buildClient(serviceUrl)
-        .withHttpHeaders(mtditidHeader, authorization)
-        .put(Json.obj())) {
-        result =>
-          result.status mustBe BAD_REQUEST
-      }
-    }
-
-    "return 400 if a downstream bad request error occurs" in new Setup {
+    "return 400 if a downstream invalid taxable entity request error occurs" in new Setup {
 
       val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
         "INVALID_TAXABLE_ENTITY_ID", "Submission has not passed validation. Invalid parameter taxableEntityId.")).toString()
 
-      stubPutWithResponseBody(desUrl, Json.toJson(fullCreateUpdateStateBenefitData).toString(), errorResponseBody, BAD_REQUEST)
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, BAD_REQUEST)
 
       authorised()
 
@@ -99,12 +82,84 @@ class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec w
       }
     }
 
-    "return 422 if a downstream invalid request error occurs" in new Setup {
+    "return 400 if a downstream invalid tax year request error occurs" in new Setup {
 
       val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
-        "INVALID_REQUEST_BEFORE_TAX_YEAR", "The remote endpoint has indicated that submission is provided before the tax year has ended.")).toString()
+        "INVALID_TAX_YEAR", "Submission has not passed validation. Invalid parameter taxYear.")).toString()
 
-      stubPutWithResponseBody(desUrl, Json.toJson(fullCreateUpdateStateBenefitData).toString(), errorResponseBody, UNPROCESSABLE_ENTITY)
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, BAD_REQUEST)
+
+      authorised()
+
+      whenReady(buildClient(serviceUrl)
+        .withHttpHeaders(mtditidHeader, authorization)
+        .put(fullCreateUpdateStateBenefitJson)) {
+        result =>
+          result.status mustBe BAD_REQUEST
+          result.body mustBe errorResponseBody
+      }
+    }
+
+    "return 400 if a downstream invalid benefit id request error occurs" in new Setup {
+
+      val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
+        "INVALID_BENEFIT_ID", "Submission has not passed validation. Invalid parameter benefitId.")).toString()
+
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, BAD_REQUEST)
+
+      authorised()
+
+      whenReady(buildClient(serviceUrl)
+        .withHttpHeaders(mtditidHeader, authorization)
+        .put(fullCreateUpdateStateBenefitJson)) {
+        result =>
+          result.status mustBe BAD_REQUEST
+          result.body mustBe errorResponseBody
+      }
+    }
+
+    "return 400 if a downstream invalid payload request error occurs" in new Setup {
+
+      val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
+        "INVALID_PAYLOAD", "Submission has not passed validation. Invalid payload.")).toString()
+
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, BAD_REQUEST)
+
+      authorised()
+
+      whenReady(buildClient(serviceUrl)
+        .withHttpHeaders(mtditidHeader, authorization)
+        .put(fullCreateUpdateStateBenefitJson)) {
+        result =>
+          result.status mustBe BAD_REQUEST
+          result.body mustBe errorResponseBody
+      }
+    }
+
+    "return 400 if a downstream invalid correlation id request error occurs" in new Setup {
+
+      val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
+        "INVALID_CORRELATIONID", "Submission has not passed validation. Invalid Header parameter CorrelationId.")).toString()
+
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, BAD_REQUEST)
+
+      authorised()
+
+      whenReady(buildClient(serviceUrl)
+        .withHttpHeaders(mtditidHeader, authorization)
+        .put(fullCreateUpdateStateBenefitJson)) {
+        result =>
+          result.status mustBe BAD_REQUEST
+          result.body mustBe errorResponseBody
+      }
+    }
+
+    "return 422 if a downstream not supported request error occurs" in new Setup {
+
+      val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
+        "NOT_SUPPORTED_TAX_YEAR", "The remote endpoint has indicated that submission is provided before the tax year has ended.")).toString()
+
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, UNPROCESSABLE_ENTITY)
 
       authorised()
 
@@ -117,12 +172,29 @@ class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec w
       }
     }
 
+    "return 403 if a downstream not forbidden request error occurs" in new Setup {
+
+      val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
+        "IGNORE_FORBIDDEN", "The remote endpoint has indicated that HMRC held State Benefit cannot be ignored.")).toString()
+
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, FORBIDDEN)
+
+      authorised()
+
+      whenReady(buildClient(serviceUrl)
+        .withHttpHeaders(mtditidHeader, authorization)
+        .put(fullCreateUpdateStateBenefitJson)) {
+        result =>
+          result.status mustBe FORBIDDEN
+          result.body mustBe errorResponseBody
+      }
+    }
+
     "return 500 if an unexpected error is returned from DES user" in new Setup {
 
-      // e,g, 404 not found is not expected as create or update will create if not found
       val errorResponseBody: String = Json.toJson(DesErrorBodyModel.parsingError).toString()
 
-      stubPutWithResponseBody(desUrl, Json.toJson(fullCreateUpdateStateBenefitData).toString(), errorResponseBody, NOT_FOUND)
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, NOT_FOUND)
 
       authorised()
 
@@ -140,7 +212,7 @@ class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec w
       val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
         "SERVICE_UNAVAILABLE", "Dependent systems are currently not responding.")).toString()
 
-      stubPutWithResponseBody(desUrl, Json.toJson(fullCreateUpdateStateBenefitData).toString(), errorResponseBody, SERVICE_UNAVAILABLE)
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, SERVICE_UNAVAILABLE)
 
       authorised()
 
@@ -159,7 +231,7 @@ class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec w
       val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
         "SERVER_ERROR", "DES is currently experiencing problems that require live service intervention.")).toString()
 
-      stubPutWithResponseBody(desUrl, Json.toJson(fullCreateUpdateStateBenefitData).toString(), errorResponseBody, INTERNAL_SERVER_ERROR)
+      stubPutWithResponseBody(desUrl, Json.toJson(fullIgnoreStateBenefit).toString(), errorResponseBody, INTERNAL_SERVER_ERROR)
 
       authorised()
 
@@ -178,7 +250,7 @@ class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec w
       unauthorisedOtherEnrolment()
 
       whenReady(buildClient(serviceUrl)
-        .withHttpHeaders(mtditidHeader)
+        .withHttpHeaders(mtditidHeader, authorization)
         .put(fullCreateUpdateStateBenefitJson)) {
         result =>
           result.status mustBe UNAUTHORIZED
@@ -195,7 +267,5 @@ class CreateUpdateStateBenefitOverrideITest extends PlaySpec with WiremockSpec w
       }
     }
 
-
   }
-
 }
