+++ it/test/connectors/GetEmploymentBenefitsConnectorISpec.scala
@@ -14,17 +14,18 @@
  * limitations under the License.
  */
 
-package connectors
+package test.connectors
 
 import com.github.tomakehurst.wiremock.http.HttpHeader
 import config.BackendAppConfig
-import connectors.getEmploymentBenefitsConnectorSpec.{expectedResponseBody, expectedResponseBodyHMRC, latestWithAllFields}
-import helpers.WiremockSpec
+import connectors.GetEmploymentBenefitsConnector
+import getEmploymentBenefitsConnectorSpec.{expectedResponseBody, expectedResponseBodyHMRC, latestWithAllFields}
 import models._
 import org.scalatestplus.play.PlaySpec
 import play.api.Configuration
 import play.api.http.Status._
 import play.api.libs.json.Json
+import test.helpers.WiremockSpec
 import uk.gov.hmrc.http.{HeaderCarrier, HeaderNames, HttpClient, SessionId}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
 import utils.DESTaxYearHelper.desTaxYearConverter
