+++ project/AppDependencies.scala
@@ -14,26 +14,25 @@
  * limitations under the License.
  */
 
-import play.core.PlayVersion.current
 import sbt.*
 
 object AppDependencies {
 
-  private val bootstrapPlay28Version = "7.23.0"
+  private val bootstrapVersion = "8.4.0"
 
   val compile: Seq[ModuleID] = Seq(
-    "uk.gov.hmrc"                  %% "bootstrap-backend-play-28" % bootstrapPlay28Version,
-    "com.fasterxml.jackson.module" %% "jackson-module-scala"      % "2.17.0"
+    "uk.gov.hmrc"                  %% "bootstrap-backend-play-30" % bootstrapVersion,
+    "com.fasterxml.jackson.module" %% "jackson-module-scala"      % "2.14.2"
   )
 
   val test: Seq[ModuleID] = Seq(
-    "uk.gov.hmrc"            %% "bootstrap-test-play-28" % bootstrapPlay28Version % Test,
-    "org.scalatest"          %% "scalatest"              % "3.2.9"                % Test,
-    "com.typesafe.play"      %% "play-test"              % current                % Test,
-    "org.pegdown"             % "pegdown"                % "1.6.0"                % "test, it",
-    "com.vladsch.flexmark"    % "flexmark-all"           % "0.36.8"               % "test, it",
-    "org.scalatestplus.play" %% "scalatestplus-play"     % "5.1.0"                % "test, it",
-    "com.github.tomakehurst"  % "wiremock-jre8"          % "3.0.1"                % "test, it",
-    "org.scalamock"          %% "scalamock"              % "6.0.0"                % Test
+    "uk.gov.hmrc"   %% "bootstrap-test-play-30" % bootstrapVersion % Test,
+    "org.scalatest" %% "scalatest"              % "3.2.15"         % Test,
+////    "com.typesafe.play"       %% "play-test"                % current                 % Test,
+    //    "org.pegdown"             %  "pegdown"                  % "1.6.0"                 % "test, it",
+    "com.vladsch.flexmark"    % "flexmark-all"             % "0.64.6" % Test,
+    "org.scalatestplus.play" %% "scalatestplus-play"       % "5.1.0"  % Test,
+    "com.github.tomakehurst"  % "wiremock-jre8-standalone" % "2.35.1" % Test,
+    "org.scalamock"          %% "scalamock"                % "5.2.0"  % Test
   )
 }
